+++ bin/ztf/archive_index.py
@@ -26,16 +26,17 @@ import argparse
 
 import pyspark.sql.functions as F
 
-from fink_broker.ztf.parser import getargs
+from fink_broker.common.spark_utils import init_sparksession, load_parquet_files
+from fink_broker.common.logging_utils import get_fink_logger, inspect_application
+from fink_broker.common.parser import getargs
 from fink_broker.common.spark_utils import ang2pix
+
 from fink_broker.ztf.hbase_utils import push_to_hbase, add_row_key
 from fink_broker.ztf.hbase_utils import assign_column_family_names
 from fink_broker.ztf.hbase_utils import load_ztf_index_cols
 from fink_broker.ztf.hbase_utils import load_ztf_crossmatch_cols
 from fink_broker.ztf.hbase_utils import select_relevant_columns
 from fink_broker.ztf.hbase_utils import bring_to_current_schema
-from fink_broker.common.spark_utils import init_sparksession, load_parquet_files
-from fink_broker.common.logging_utils import get_fink_logger, inspect_application
 
 from fink_filters.classification import extract_fink_classification
 from fink_utils.spark.utils import check_status_last_prv_candidates
