+++ bin/ztf/archive_low_state_blazar.py
@@ -24,7 +24,7 @@ import os
 import numpy as np
 from astropy.time import Time
 
-from fink_broker.ztf.parser import getargs
+from fink_broker.common.parser import getargs
 from fink_broker.common.spark_utils import init_sparksession, load_parquet_files
 from fink_broker.common.logging_utils import get_fink_logger, inspect_application
 from fink_broker.ztf.hbase_utils import push_full_df_to_hbase
@@ -32,7 +32,7 @@ from fink_broker.ztf.hbase_utils import push_full_df_to_hbase
 from fink_filters.filter_blazar_low_state.filter import low_state_filter
 from fink_filters.filter_blazar_new_low_state.filter import new_low_state_filter
 
-from fink_science.standardized_flux.processor import standardized_flux
+from fink_science.ztf.standardized_flux.processor import standardized_flux
 
 from fink_utils.spark.utils import concat_col
 from fink_utils.tg_bot.utils import send_simple_text_tg
