+++ bin/ztf/archive_science.py
@@ -24,13 +24,11 @@
 
 import argparse
 
-from fink_broker.ztf.parser import getargs
+from fink_broker.common.parser import getargs
 from fink_broker.common.spark_utils import init_sparksession, load_parquet_files
-
-from fink_broker.ztf.hbase_utils import push_full_df_to_hbase
 from fink_broker.common.spark_utils import list_hdfs_files
-
 from fink_broker.common.logging_utils import get_fink_logger, inspect_application
+from fink_broker.ztf.hbase_utils import push_full_df_to_hbase
 
 
 def main():
