+++ bin/ztf/archive_sso_cand.py
@@ -27,12 +27,11 @@ import os
 import argparse
 import pandas as pd
 
-from fink_broker.ztf.parser import getargs
+from fink_broker.common.logging_utils import get_fink_logger, inspect_application
+from fink_broker.common.parser import getargs
 from fink_broker.common.spark_utils import init_sparksession
 
 from fink_broker.ztf.hbase_utils import push_to_hbase
-
-from fink_broker.common.logging_utils import get_fink_logger, inspect_application
 
 
 def main():
