+++ bin/ztf/archive_sso_resolver.py
@@ -23,8 +23,8 @@ import pandas as pd
 import numpy as np
 
 from fink_broker.common.spark_utils import init_sparksession
-from fink_broker.ztf.hbase_utils import push_to_hbase
 from fink_broker.common.logging_utils import get_fink_logger, inspect_application
+from fink_broker.ztf.hbase_utils import push_to_hbase
 
 from fink_spins.ssoft import rockify
