+++ bin/ztf/archive_statistics.py
@@ -23,9 +23,9 @@ import pandas as pd
 import pyspark.sql.functions as F
 
 from fink_broker.common.spark_utils import init_sparksession
-from fink_broker.ztf.hbase_utils import push_to_hbase
-from fink_broker.ztf.parser import getargs
+from fink_broker.common.parser import getargs
 from fink_broker.common.logging_utils import get_fink_logger, inspect_application
+from fink_broker.ztf.hbase_utils import push_to_hbase
 
 from fink_filters.classification import extract_fink_classification
 from fink_filters.filter_simbad_candidates.filter import simbad_candidates
