+++ bin/ztf/generate_ssoft.py
@@ -21,7 +21,7 @@ import datetime
 from fink_broker.common.logging_utils import get_fink_logger, inspect_application
 from fink_broker.common.spark_utils import init_sparksession
 
-from fink_science.ssoft.processor import build_the_ssoft
+from fink_science.ztf.ssoft.processor import build_the_ssoft
 
 
 def main():
