+++ bin/ztf/save_distribution_schema.py
@@ -21,7 +21,7 @@ import subprocess
 import glob
 import json
 
-from fink_broker.ztf.parser import getargs
+from fink_broker.common.parser import getargs
 from fink_broker.common.spark_utils import init_sparksession, load_parquet_files
 from fink_broker.common.avro_utils import readschemafromavrofile
 from fink_broker.common.logging_utils import get_fink_logger, inspect_application
