+++ bin/ztf/tns_resolver.py
@@ -22,7 +22,7 @@ import os
 from fink_tns.utils import download_catalog
 from fink_broker.common.spark_utils import init_sparksession
 from fink_broker.ztf.hbase_utils import add_row_key, push_to_hbase
-from fink_broker.ztf.parser import getargs
+from fink_broker.common.parser import getargs
 
 
 def format_tns_for_hbase(pdf: pd.DataFrame) -> pd.DataFrame:
